refactor: move chat pages into features with stubs
